better assistant message part UX, error handling. file organisation, cleanup
