+++ apps/petrinaut-website/package.json
@@ -8,7 +8,8 @@
     "dev": "vite",
     "fix:eslint": "oxlint --fix --type-aware --report-unused-disable-directives-severity=error .",
     "lint:eslint": "oxlint --type-aware --report-unused-disable-directives-severity=error .",
-    "lint:tsc": "tsgo --noEmit"
+    "lint:tsc": "tsgo --noEmit",
+    "preview": "vite preview"
   },
   "dependencies": {
     "@ai-sdk/openai": "3.0.63",
