+++ libs/@hashintel/petrinaut/src/ui/views/Editor/panels/ai-assistant-panel/ai-assistant-contents/shared/collapsible-content-style.ts
@@ -1,0 +1,13 @@
+import { css } from "@hashintel/ds-helpers/css";
+
+export const collapsibleContentStyle = css({
+  overflow: "hidden",
+  animationDuration: "[200ms]",
+  animationTimingFunction: "ease-in-out",
+  "&[data-state=open]": {
+    animationName: "expand",
+  },
+  "&[data-state=closed]": {
+    animationName: "collapse",
+  },
+});
